Move RentalInventory to @/data
